Updating motor control
It's functional. Currently pixel/empty is mapped to sw[2]. Should be simple enough to change. Only other thingI can think of is adding a flag once a pixel/empty has been completed. Then the top level will send in the next pixel value.

Might need to play around with the PWMing a bit. Currently it takes 1:15 to draw a line of empty pixels. There are about 106 lines, which means that it'll take like 2 hours minimum to draw a whole picture?? THAT'S TERRIBLE